Remove juggler from peerDependencies.
As of #9, loopback-testing no longer `require()s` loopback directly.
Module users have to supply their loopback models by passing the
constructor to helper functions.

The reason for keeping loopback in peerDependencies is to control
which loopback version are compatible with loopback-testing.
To achieve that, loopback-datasource-juggler does not need to be
included in peerDependencies.